# openapi.yaml
openapi: 3.1.1
info:
  title: E-commerce Simulation Agent API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
  summary: API для агента, управляющего интернет-магазином в симуляции
  description: |
    Публичный HTTP API для запуска детерминированной симуляции и управления
    интернет-магазином от имени агента.

    В симуляции есть три типа страниц:

    1. `product_list` — список товаров;
    2. `product_page` — страница конкретного товара;
    3. `purchase` — страница покупки. Успешный вызов этой страницы означает покупку.

    Каждый запрос посетителя создаёт нагрузку в условных единицах на заданное
    время и оставляет запись в site logs со статусом 200 или 500. Серверы имеют
    ограниченную capacity и стоят денег. Агент может наблюдать метрики и логи,
    изменять число серверов, применять fixes, запускать подготовленные deployments,
    выполнять deployments строго по очереди и явно прокручивать время.

    `seed < 0` обозначает заранее созданный вручную мир. Он обязан существовать в
    базе данных. `seed > 0` обозначает генерируемый мир: сервер использует уже
    сохранённый мир с тем же seed и версией генератора либо детерминированно создаёт
    и сохраняет его. `seed = 0` запрещён.

    Seed идентифицирует мир, а `run_id` — отдельную попытку агента пройти этот мир.
    Каждый новый успешный `/start` создаёт независимый run, кроме идемпотентного
    повтора с теми же `agent_id`, `agent_version` и `request_id`.

...

...
servers:
...
security: []
...
paths:
...
  /v1/runs/{run_id}/fixes:
    post:
      tags: [Fixes]
      operationId: applyFix
      summary: Применить fix по тексту из site logs
      description: |
        Если `message` совпадает с fix-текстом активного бага или DDoS-события,
        требующего mitigation, причина ошибки полностью деактивируется. Повтор с
        тем же `request_id` возвращает исходный результат.
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyFixRequest'
            example:
              request_id: fix-20260824-001
              message: FIX-7F92A
      responses:
        '200':
          description: Fix проверен и, при совпадении, применён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyFixResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/runs/{run_id}/deployments:
    get:
      tags: [Deployments]
      operationId: getDeployments
      summary: Получить упорядоченную очередь deployments
      description: |
        Deployments являются последовательными задачами агента. Запустить можно
        только первый незавершённый deployment со статусом `available`; последующие
        имеют статус `locked`. Скрытые эффекты и будущие баги не раскрываются.
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Список deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentsResponse'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      tags: [Deployments]
      operationId: startDeployment
      summary: Запустить подготовленный deployment
      description: |
        Deployment может изменить нагрузку страниц, время удержания ресурсов,
        вероятность ошибок или активировать/деактивировать баги. В запросе передаётся
        только ID подготовленного deployment, а не произвольный код. Запрос
        deployment не по очереди отклоняется с `DEPLOYMENT_ORDER_VIOLATION`.
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartDeploymentRequest'
            example:
              request_id: deploy-20260824-001
              deployment_id: catalog-cache-v1
      responses:
        '202':
          description: Deployment принят к выполнению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAcceptedResponse'
        '200':
          description: Результат идемпотентного повтора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAcceptedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          description: Run или deployment не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/runs/{run_id}/time/advance:
    post:
      tags: [Time]
      operationId: advanceTime
      summary: Явно прокрутить симуляционное время
      description: |
        `duration_seconds` должен быть не меньше 300. Фактическое продвижение равно
        `max(real_elapsed_since_previous_request, duration_seconds)`. Повтор с тем же
        `request_id` не продвигает время второй раз.
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvanceTimeRequest'
            example:
              request_id: advance-20260824-001
              duration_seconds: 86400
      responses:
        '200':
          description: Время продвинуто, наступившие события обработаны
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvanceTimeResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      description: Неподбираемый буквенно-цифровой идентификатор запуска
      schema:
        $ref: '#/components/schemas/RunId'
...
  responses:
    InvalidRequest:
      description: Запрос не соответствует контракту или нарушает ограничения
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidSeed:
              value:
                error: INVALID_SEED
                message: Seed must be a non-zero integer
            minimumAdvance:
              value:
                error: MINIMUM_ADVANCE_IS_300_SECONDS
                message: duration_seconds must be at least 300
    RunNotFound:
      description: Run с таким ID не существует
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: RUN_NOT_FOUND
            message: Run does not exist
    Conflict:
      description: Конфликт состояния, конкурентный запрос или недопустимый порядок действий
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            concurrentRequest:
              value:
                error: CONCURRENT_RUN_REQUEST
                message: Another request is currently modifying this run
            taskOrder:
              value:
                error: DEPLOYMENT_ORDER_VIOLATION
                message: The requested deployment is not currently available
  schemas:
    RunId:
      type: string
      description: Криптографически случайный base62 ID; не является хешем seed
      pattern: '^[A-Za-z0-9]{20,64}$'
      examples: [7Kp3mN8qR2vX5cL9tF4a]
    RequestId:
      type: string
      description: Идемпотентный ID, уникальный в пределах агента или run
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      examples: [request-20260824-001]
...
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    Clock:
      type: object
      additionalProperties: false
      required:
        - simulation_time
        - simulation_ends_at
        - remaining_seconds
        - real_elapsed_seconds
        - applied_advance_seconds
      properties:
        simulation_time:
          type: string
          format: date-time
          description: Текущее симуляционное время после автоматического продвижения
        simulation_ends_at:
          type: string
          format: date-time
          description: Максимальный момент времени, до которого работает этот run
        remaining_seconds:
          type: number
          minimum: 0
          description: Сколько симуляционных секунд осталось до завершения run
        real_elapsed_seconds:
          type: number
          minimum: 0
          description: Реальное время с предыдущего вызова этого run
        applied_advance_seconds:
          type: number
          minimum: 0
          description: Фактическое продвижение симуляции перед выполнением запроса
...
    OperationAcceptedResponse:
      type: object
      additionalProperties: false
      required: [clock, operation_id, status]
      properties:
        clock:
          $ref: '#/components/schemas/Clock'
        operation_id:
          type: string
        status:
          $ref: '#/components/schemas/OperationStatus'
        estimated_complete_at:
          type: [string, 'null']
          format: date-time
    ApplyFixRequest:
      type: object
      additionalProperties: false
      required: [request_id, message]
      properties:
        request_id:
          $ref: '#/components/schemas/RequestId'
        message:
          type: string
          minLength: 1
          maxLength: 4096
    ApplyFixResponse:
      type: object
      additionalProperties: false
      required: [clock, applied]
      properties:
        clock:
          $ref: '#/components/schemas/Clock'
        applied:
          type: boolean
        fixed_bug:
          type: [string, 'null']
          description: Публичный ID исправленного бага или null при несовпадении
        fixed_attack:
          type: [string, 'null']
          description: Публичный ID остановленной атаки или null
        message:
          type: string
    DeploymentStatus:
      type: string
      description: |
        `locked` — предыдущий deployment ещё не завершён; `available` — текущая
        задача, которую агент может запустить.
      enum: [locked, available, queued, running, succeeded, failed]
    Deployment:
      type: object
      additionalProperties: false
      required: [deployment_id, sequence, name, status]
      properties:
        deployment_id:
          type: string
        sequence:
          type: integer
          minimum: 1
          description: Позиция deployment в очереди задач
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/DeploymentStatus'
        operation_id:
          type: [string, 'null']
    DeploymentsResponse:
      type: object
      additionalProperties: false
      required: [clock, deployments]
      properties:
        clock:
          $ref: '#/components/schemas/Clock'
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/Deployment'
    StartDeploymentRequest:
      type: object
      additionalProperties: false
      required: [request_id, deployment_id]
      properties:
        request_id:
          $ref: '#/components/schemas/RequestId'
        deployment_id:
          type: string
          minLength: 1
          maxLength: 128
...
    AdvanceTimeRequest:
      type: object
      additionalProperties: false
      required: [request_id, duration_seconds]
      properties:
        request_id:
          $ref: '#/components/schemas/RequestId'
        duration_seconds:
          type: integer
          minimum: 300
          description: Запрошенная прокрутка; минимум пять минут
    AdvanceTimeResponse:
      type: object
      additionalProperties: false
      required:
        - clock
        - previous_simulation_time
        - requested_duration_seconds
        - processed_events
        - new_logs
      properties:
        clock:
          $ref: '#/components/schemas/Clock'
        previous_simulation_time:
          type: string
          format: date-time
        requested_duration_seconds:
          type: integer
          minimum: 300
        processed_events:
          type: integer
          minimum: 0
        new_logs:
          type: integer
          minimum: 0
        logs_cursor:
          type: [string, 'null']
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          examples: [RUN_NOT_FOUND]
        message:
          type: string
        details:
          type: object
          additionalProperties: true